// spartak.web/src/api/dto/IFanCard.ts
import { LocaleType } from "./LocaleType";

export interface IFanCard {
  title: LocaleType;
  text: LocaleType;
  subTitle1: LocaleType;
  text1: LocaleType;
  subTitle2: LocaleType;
  text2: LocaleType;
  list: LocaleType[];
  text2_1: LocaleType;
  subTitle3: LocaleType;
  text3: LocaleType;
  text4: LocaleType;
}